Add files via upload
This dashboard, crafted with Microsoft Excel for data cleaning and Power BI for transformation, analysis, and visualization, offers insightful metrics to gauge the impact and effectiveness of our placement initiatives.

Visualizing placements by sector and gender through pie charts allows for a quick assessment of the distribution of opportunities across different industries and the gender inclusivity in job placements.

The line chart provides a quarterly trend analysis, offering stakeholders a dynamic view of the placement progress throughout the year. This enables the identification of seasonal patterns or emerging trends in youth employment.

The bar chart breaks down placements by region, age, and level of education, providing detailed insights into the demographics of successful placements. This information is invaluable for tailoring future initiatives to specific regional and demographic needs.

The doughnut chart visualizes placements by industries, offering a holistic view of the diverse sectors where youth are finding employment. This information aids in understanding the breadth and depth of our impact on the workforce.

Key metrics are summarized in cards, including the total number of placements, the distribution of placements by gender, and disability status. These cards provide a quick snapshot of the overall success and demographic inclusivity of our placement initiatives.